pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "openai>=1.97.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
